requirements.txt
flask==3.1.0
dnspython==2.4.2